requirements.txt
grpcio==1.44.0
grpcio-tools==1.44.0
grpcio-reflection==1.44.0
grpcio-health-checking==1.44.0
grpcio-testing==1.44.0

pandas==1.3.2
joblib==1.0.1
numpy~=1.21.5
scikit-learn==1.0.2